Lazily register more cudf functions and move to backends file (#4396)
* Add has_parallel_type

This is a bit nicer than `isinstance(obj, parallel_types())` because it
works also with lazily registered types.

* Add more cudf lazy registrations

* Add explicit backends.py files in dask.array/dataframe

This collects lazily registered functions for cupy, sparse, scipy, and
cudf.
