+++ dask/array/__init__.py
@@ -47,7 +47,7 @@ try:
         from .reductions import nanprod, nancumprod, nancumsum
     with ignoring(ImportError):
         from . import ma
-    from . import random, linalg, overlap, fft
+    from . import random, linalg, overlap, fft, backends
     from .overlap import map_overlap
     from .wrap import ones, zeros, empty, full
     from .creation import ones_like, zeros_like, empty_like, full_like
